refactor : Exclude 사용하여 타입 재정의 및 범용적이지 않은 타입 위치 조정
